apis/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
clickhouse-driver>=0.2.6
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
loguru>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5 